Cambios en pantalla,  estructura de archivos
