+++ src/screens/LoginAdministrador/index.js
@@ -1,0 +1,3 @@
+import LoginAdministrator from './LoginAdministrator'
+
+export default LoginAdministrator
